Fix vehicle destroy sync (except cyclops)
