+++ NitroxClient/Communication/Packets/Processors/VehicleDestroyedProcessor.cs
@@ -3,27 +3,24 @@ using NitroxClient.Communication.Packets.Processors.Abstract;
 using NitroxClient.GameLogic;
 using NitroxModel.Packets;
 
-namespace NitroxClient.Communication.Packets.Processors
+namespace NitroxClient.Communication.Packets.Processors;
+
+public class VehicleDestroyedProcessor : ClientPacketProcessor<VehicleDestroyed>
 {
-    public class VehicleDestroyedProcessor : ClientPacketProcessor<VehicleDestroyed>
+    private readonly IPacketSender packetSender;
+    private readonly Vehicles vehicles;
+
+    public VehicleDestroyedProcessor(IPacketSender packetSender, Vehicles vehicles)
     {
-        private readonly IPacketSender packetSender;
-        private readonly Vehicles vehicles;
-        private readonly PlayerManager remotePlayerManager;
+        this.packetSender = packetSender;
+        this.vehicles = vehicles;
+    }
 
-        public VehicleDestroyedProcessor(IPacketSender packetSender, Vehicles vehicles, PlayerManager remotePlayerManager)
+    public override void Process(VehicleDestroyed packet)
+    {
+        using (packetSender.Suppress<VehicleDestroyed>())
         {
-            this.packetSender = packetSender;
-            this.vehicles = vehicles;
-            this.remotePlayerManager = remotePlayerManager;
-        }
-
-        public override void Process(VehicleDestroyed packet)
-        {
-            using (packetSender.Suppress<VehicleDestroyed>())
-            {
-                vehicles.DestroyVehicle(packet.Id, packet.GetPilotingMode);
-            }
+            vehicles.DestroyVehicle(packet.Id);
         }
     }
 }
